+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [UNRELEASED] neptune-client 0.16.3
+## neptune-client 0.16.3
 
 ## Features
 - Added fetching Models method to Project ([#916](https://github.com/neptune-ai/neptune-client/pull/916))
@@ -7,6 +7,9 @@
 - Fix computing of a multipart upload chunk size ([#897](https://github.com/neptune-ai/neptune-client/pull/897))
 - Matching all listed tags instead of any when calling `fetch_runs_table` ([#899](https://github.com/neptune-ai/neptune-client/pull/899))
 - Fix invalid processing of delete followed by file upload in a single batch ([#880](https://github.com/neptune-ai/neptune-client/pull/880))
+
+## Changes
+- `click.echo` replaced with `logging` ([#903](https://github.com/neptune-ai/neptune-client/pull/903))
 
 ## neptune-client 0.16.2